Replace old files with new structure
